refactor: Date 파일 위치 변경에 따른 테스트코드 수정
